Remove COPYING2
COPYING2 was an attempt in 2007 to Affero-ize GPLv2 in a cleaner manner than the AGPLv1
did it. A Debian maitainer objected to the file and marked the project for deletion
from Debian's repository. Because his objections to the file were primarily ad hominem,
the only way to satisfy the maintainer was to delete the file. So it goes.